Documentation updates (STR #505, STR #513)
The old source file "fl_set_gray.cxx" is not needed (STR #516)

Fl_Text_Display still called delete[] instead of free() in one
place (STR #503)

Fl_File_Chooser didn't correctly call isprint() and isspace()
when checking to see if the current file was text that can be
previewed (STR #517)

FLUID didn't compile with Borland C++ due to a compiler bug (STR
#496)

Fl_Positioner did not handle reversed min and max values (STR
#510)

fl_descent(), fl_height(), and fl_width() would crash a program
if you didn't call fl_font() first; they now return -1 if no
font is set (STR #500)


git-svn-id: file:///fltk/svn/fltk/branches/branch-1.1@3805 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ fluid/fluid.cxx
@@ -1,5 +1,5 @@
 //
-// "$Id: fluid.cxx,v 1.15.2.13.2.44 2004/04/11 04:38:56 easysw Exp $"
+// "$Id: fluid.cxx,v 1.15.2.13.2.45 2004/09/07 20:59:16 easysw Exp $"
 //
 // FLUID main entry for the Fast Light Tool Kit (FLTK).
 //
@@ -135,7 +135,7 @@ void save_position(Fl_Window *w, const char *prefsName) {
   pos.set("y", w->y());
   pos.set("w", w->w());
   pos.set("h", w->h());
-  pos.set("visible", w->shown() && w->visible() );
+  pos.set("visible", (int)(w->shown() && w->visible()));
 }
 
 Fl_Window *main_window;
@@ -886,5 +886,5 @@ int main(int argc,char **argv) {
 }
 
 //
-// End of "$Id: fluid.cxx,v 1.15.2.13.2.44 2004/04/11 04:38:56 easysw Exp $".
+// End of "$Id: fluid.cxx,v 1.15.2.13.2.45 2004/09/07 20:59:16 easysw Exp $".
 //
